+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 # [0.1.0] - 2020-2-16
 ### Added
- - rough RNN code to generate Shakespeare in `shakespeare_ai`
+ - `prepare.py` code to input a training dataset and prepare it for modeling
+ - `model.py` code to define a PyTorch RNN
+ - `learner.py` wrapper over `model.py` that nicely handles training, predicting, saving, and loading a model
 
 # [0.0.0] - 2020-2-16
 ### Added